pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "markdown-rag"
version = "0.1.0"
description = "Semantic search and QA system for internal Markdown documents"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "chromadb>=1.5",
    "sentence-transformers>=5.0",
    "markdown-it-py>=4.0",
    "pydantic>=2.12",
    "pydantic-settings>=2.13",
    "openai>=2.29",
    "httpx>=0.28",
    "python-frontmatter>=1.1",
    "fastapi>=0.135",
    "uvicorn>=0.42",
    "llama-cpp-python>=0.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0",
    "pytest-cov>=7.1",
    "pytest-asyncio>=1.0",
    "httpx>=0.28",
    "ruff>=0.15",
]

[project.scripts]
mdrag = "markdown_rag.cli.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/markdown_rag"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.coverage.run]
source = ["markdown_rag"]

[tool.coverage.report]
fail_under = 85
show_missing = true
